Code/backend/requirements.txt
Flask
Flask-JWT-Extended
psycopg2-binary
python-dotenv
Flask-cors
Flask-SQLAlchemy